# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Rep API",
        "description": "E-Rep Pharmaceutical Platform API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/doctor/messages": {
            "get": {
                "tags": [
                    "Doctor - Messages"
                ],
                "summary": "List messages",
                "operationId": "97f7ba6199634587eeb632e67564ebdb",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Doctor - Messages"
                ],
                "summary": "Send message to medical rep",
                "operationId": "0e3c0aa74690658cf8334a2ec43084c4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "receiver_rep_id": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/messages/{id}/read": {
            "post": {
                "tags": [
                    "Doctor - Messages"
                ],
                "summary": "Mark message as read",
                "operationId": "b16ed8d25bb034a37d7567be4b74ada6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/notifications": {
            "get": {
                "tags": [
                    "Doctor - Messages"
                ],
                "summary": "List notifications",
                "operationId": "b6724e540f93c32f1989407d4ff8b951",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/notifications/read-all": {
            "post": {
                "tags": [
                    "Doctor - Messages"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "f56519fb0491205c7fafec21150c85d0",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/doctor/report/generate": {
            "get": {
                "tags": [
                    "Doctor - Report"
                ],
                "summary": "Download PDF activity report",
                "operationId": "6985a84120c461961712bbe32ebf951c",
                "responses": {
                    "200": {
                        "description": "PDF binary stream",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/rewards": {
            "get": {
                "tags": [
                    "Doctor - Points"
                ],
                "summary": "List rewards with can_redeem flag",
                "operationId": "82d59ec7a7c48678495135e444c08b9f",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/redemptions": {
            "get": {
                "tags": [
                    "Doctor - Points"
                ],
                "summary": "List my redemptions",
                "operationId": "81a24c7150c6862ef215c8b415678e16",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/doctor/rewards/{rewardId}/redeem": {
            "post": {
                "tags": [
                    "Doctor - Points"
                ],
                "summary": "Redeem reward",
                "operationId": "dada293538506174ecc813772dbaf674",
                "parameters": [
                    {
                        "name": "rewardId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Not enough points"
                    },
                    "404": {
                        "description": "Reward not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}